requirements.txt
opencv-python
scikit-learn
pillow
numpy
